+++ server/src/main/java/application/model/AuthorizationDetails.java
@@ -2,9 +2,9 @@ package application.model;
 
 import application.entity.Balance;
 import application.entity.Client;
+import application.exception.ClientNotFoundException;
 import dto.ClientRequestOperations;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import org.springframework.stereotype.Service;
 import view.Currencies;
 
@@ -30,7 +30,8 @@ public class AuthorizationDetails {
         currencies = new ArrayList<>();
         Set<Balance> balances = client.getBalance();
         currentBalance = balances.stream()
-                .filter(b -> b.getName().equals(request.getCurrency())).findAny().get();
+                .filter(b -> b.getName().equals(request.getCurrency())).findAny()
+                .orElseThrow(() -> new ClientNotFoundException("Not found balance!"));
         return this;
     }
 }
