+++ server/src/main/java/application/security/filters/JwtRequestFilter.java
@@ -16,7 +16,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-
 @Component
 @AllArgsConstructor
 public class JwtRequestFilter extends OncePerRequestFilter {
